rename class to code scanner search
