+++ src/utils/codeScannerSearch.ts
